Move seed to jobs.
