+++ src/main/kotlin/uk/gov/justice/digital/hmpps/approvedpremisesapi/seed/SeedService.kt
@@ -34,13 +34,13 @@ import uk.gov.justice.digital.hmpps.approvedpremisesapi.cas1.seed.Cas1UpdatePrem
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.cas1.seed.Cas1UpdateSpaceBookingSeedJob
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.cas1.seed.Cas1UsersSeedJob
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.cas1.seed.Cas1WithdrawPlacementRequestSeedJob
-import uk.gov.justice.digital.hmpps.approvedpremisesapi.cas2hdc.seed.Cas2ApplicationsSeedJob
-import uk.gov.justice.digital.hmpps.approvedpremisesapi.cas2hdc.seed.Cas2NomisUserEmailSeedJob
-import uk.gov.justice.digital.hmpps.approvedpremisesapi.cas2hdc.seed.Cas2UpdateAssessmentStatusSeedJob
-import uk.gov.justice.digital.hmpps.approvedpremisesapi.cas2hdc.seed.Cas2UsersSeedJob
-import uk.gov.justice.digital.hmpps.approvedpremisesapi.cas2hdc.seed.ExternalUsersSeedJob
-import uk.gov.justice.digital.hmpps.approvedpremisesapi.cas2hdc.seed.NomisUsersSeedJob
-import uk.gov.justice.digital.hmpps.approvedpremisesapi.cas2hdc.seed.ShortTermAccommodationCreateOmusSeedJob
+import uk.gov.justice.digital.hmpps.approvedpremisesapi.cas2hdc.jobs.seed.Cas2ApplicationsSeedJob
+import uk.gov.justice.digital.hmpps.approvedpremisesapi.cas2hdc.jobs.seed.Cas2NomisUserEmailSeedJob
+import uk.gov.justice.digital.hmpps.approvedpremisesapi.cas2hdc.jobs.seed.Cas2UpdateAssessmentStatusSeedJob
+import uk.gov.justice.digital.hmpps.approvedpremisesapi.cas2hdc.jobs.seed.Cas2UsersSeedJob
+import uk.gov.justice.digital.hmpps.approvedpremisesapi.cas2hdc.jobs.seed.ExternalUsersSeedJob
+import uk.gov.justice.digital.hmpps.approvedpremisesapi.cas2hdc.jobs.seed.NomisUsersSeedJob
+import uk.gov.justice.digital.hmpps.approvedpremisesapi.cas2hdc.jobs.seed.ShortTermAccommodationCreateOmusSeedJob
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.cas2v2.seed.Cas2v2ApplicationsSeedJob
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.cas2v2.seed.Cas2v2UsersSeedJob
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.cas3.seed.Cas3AssignApplicationToPduSeedJob
